+++ CHANGELOG.md
@@ -4,6 +4,8 @@ UNRELEASED
 
 - Send gitaly-ruby exceptions to their own DSN
   https://gitlab.com/gitlab-org/gitaly/merge_requests/656
+- Run Go test suite with '-race' in CI
+  https://gitlab.com/gitlab-org/gitaly/merge_requests/654
 - Implement Get{Tag,Commit}Messages RPCs
   https://gitlab.com/gitlab-org/gitaly/merge_requests/646
 - Fix directory permission walker for Go 1.10